scripts/requirements.txt
fastapi
uvicorn
openai
replicate
pillow
